Tweak insecure drop user clean up (#52855)
Adjusts the logging that occurs when deleting inactive temporary
host users to be less spammy. Some logs were moved from INFO to
DEBUG and others were updated to only occur if cleanup takes a
particular action.

The periodic clean up loop was also updated to no longer exit on
a NotFoundError. It seemed like a more robust solution to keep
failing and report the problem than receive a single NotFoundError
and give up on cleaning up users entirely until the process was
restarted.