Machine ID: Update Kubernetes access guide for kubernetes/v2 (#52830)
* Machine ID: Document `kubernetes/v2` service

This adds initial documentation for the new `kubernetes/v2` service.

This only includes reference documentation for now, since this
service was added partway through v17; the existing Kubernetes guide
will be updated in a separate PR for v18 and won't be backported to
v17.

Partially fixes #52412

* Review feedback, and add an FAQ entry about the new service

* Add more detailed version note

* Fix broken link

* Machine ID: Update Kubernetes access guide for kubernetes/v2

This updates the Kubernetes access guide to use the new
`kubernetes/v2` service which allows access to multiple clusters
from one output.

As support for this came out partway through v17 (v17.2.7), this guide
update should not be backported to v17.

Fixes #52412