Add a tab-by-tab mode for creating new roles (#52859)
According to the UI designs, the user needs to be walked through every
tab specifically when creating a new role (as opposed to modifying an
existing one). In this mode, validation is triggered every time the user
wants to progress to the next tab. Tab switching is allowed for tabs
that have already been visited.