Allow IC plugin to import groups multiple times (#6129)
* Allow IC plugin to import groups multiple times

The IC plugin automatically imports IC groups into Teleport on its
first run, and uses the plugin status block to mark the import as complete,
and is not required the next time the plugin starts up.

New functionality in `tctl` allows Teleport admins to edit the IC plugin
resource in order to modify group, account and permission set filters.
At the moment, the group filters are only applied during this initial import,
so changing them has no effect. In order to apply the new group filters, we need
to re-run the import process.

Prior to this patch, plugin resources updated via `tctl edit` had any changes
to their status blocks automatically reverted during the plugin update. This is
by design in order to stop users from modifying the plugin status in a way that
might break a running plugin. But it also leaves no avenue for signalling to the
IC plugin that the import process should be re-run.

This patch adds a special case to the plugin resource validation that allows an
update to modify the pligin status if (and only if) the p;ugin resource is for
the IC plugin AND the Group Import State is requesting that the import process
gets re-run.

* Add re-import test, fix imports, etc

* Unused parameter removal

* Make a group filter change trigger a re-import

* Update lib/plugins/pluginsv1/service.go

Co-authored-by: Marek Smoliński <marek@goteleport.com>

* Test cleanup

---------

Co-authored-by: Marek Smoliński <marek@goteleport.com>

Source-commit: gravitational/teleport.e@f4014696fd59f11e6145a263bd659af8cc0abfa6
Source-pull-requests: gravitational/teleport.e#6129