docs: Update preflight docs for building buildboxes (#50752)
Update the preflight docs to provide instructions for building the
buildboxes for the new version that master becomes. The buildboxes build
on a schedule and that can mean up to a week will need to pass before
the buildboxes are automatically built, given that the assets buildbox
is build only once a week (Sundays) and that is needed to build the
centos7 buildboxes. The updated instructions describe how to kick off
that build automatically.