Always use the same address format in session trackers
It is assumed that the `target_address` of a SessionTracker is the
host UUID for SSH sessions. However, this was only the case when in
node recording mode. The value is derived from the `ServerID` in the
return value from `lib/srv.Server.TargetMetadata`. In node recording
mode the value is populated from the host uuid of the process running
the SSH service. In proxy recording mode, the value is populated
from the server id in the dial request. The discrepency arises from
the fact that the router was altering the server id to include the
cluster name suffix to populate the prinicpals appropriately.

To remedy this some slight tweaks were made to the semantics of the
srv.Server.HostUUID and the reversetunnelclient.DialParams. All
dial requests for SSH targets now provide the resolved types.Server.
This server is used as the source of truth for the hostname and
host uuid of the target by the forward server. `(forward.Server) ID()`
now returns the proxy host uuid instead of a uuid generated when
creating the server. `(forward.Server) HostUUID()` now returns the
host uuid of the target server.

Fixes #42346.