[v15] Remove v17 and v16 changelog entries on branch/v15
Remove all the v17 and v16 entries in the CHANGELOG.md file for
branch/v15. This puts the latest v15 changelog at the top of the file
which works better on the docs page for the changelog. Also the v17 and
v16 entries were getting stale as we do not update all the branches'
changelogs when we do a v17 or v16 release.