[v16] Remove v17 changelog entries on branch/v16
Remove all the v17 entries in the CHANGELOG.md file for branch/v16. This
puts the latest v16 changelog at the top of the file which works better
on the docs page for the changelog. Also the v17 entries were getting
stale as we do not update all the branches' changelogs when we do a v17
release.