[v17] ci: Remove changelog check concurrency (#52944)
Remove the changelog.yaml workflow concurrency check that ensures only
one instance runs at a time, with later runs killing earlier ones. This
was previously used to ensure we did not put multiple messages on the
PR, but as we no longer write messages to the PR, we don't need this.

The concurrency constraint causes an occasional problem where due to
workflow scheduling irregularities an earlier workflow run starts after
a later one, killing it and leaving the PR without a required check
having run on the latest commit.