Respond to zmb3 feedback
Create a Request Access to Resources guide that shows end-users how to
create Access Requests. As a first iteration, this guide uses only
information moved from the Resource Requests guide. Instead of manually
adding an entry to the User Guides section in `sidebar.json`, this change
configures the sidebar to generate entries for this section
automatically.