Move version support table to Upcoming Releases
Ports #53003 to `master`

This change helps guarantee that we'll update the version support table
at the same time as the upcoming release table to keep our support
windows up to date. Since we otherwise do not update the two tables at
the same time, the supported version table has become out of date.

This change also updates support windows based on the plan to release
v18 in May, 2025, with the assumption that we release a major version
every 4 months.

This change also moves the compatibility description from FAQ to
Upcoming Releases so we keep this info in one place, and links to
Upcoming Releases from the relevant section of the FAQ.

Since the docs engine automatically copies the Upcoming Releases page
from the default version to all four versions of the docs, this change
does not edit the Upcoming Releases page for this version.

This change removes an item from the docs test plan re: editing the
supported versions table, since we'll be doing this continuously rather
than when we release a new major version.