[v17] lib/utils/typical: support `==` and `!=` operators for integers (#52991)
* lib/utils/typical: support `==` and `!=` operators for integers

In #52460, we're adopting the Teleport predicate language for Workload Identity
rules. We currently allow you to match on the equality/inequality of the attested
process' PID, GID, and UID, but this wasn't supported in the predicate language.

We now check the operand types at compile-time and choose the correct
booleanOperator. If both operands are dynamically typed (e.g. `ifelse`) we do
not allow you to use the `==` or `!=` operators.

* Fix compile/parse time nomenclature

* Rename type parameters for clarity