[v17] Reorganize the feature matrix (#53167)
Backports #52641

* Reorganize the feature matrix

This change organizes the feature matrix according to the latest product
categorization scheme, and moves the feature matrix into its own guide
to make it more discoverable. While there is opportunity to rework the
rows within the feature matrix, this change minimizes the scope of the
first iteration by reorganizing the existing rows while leaving them for
the most part unchanged.

In the new feature matrix page, each Teleport product has its own
H2-level section. Product descriptions come from official Marketing
language, focusing on concrete statements about what each product
comprises and how it works and leaving out statements of value (which
are not typically found in documentation). This change does not include
sections for categories with no rows in the current table.

* Respond to zmb3 feedback

Add minor wording tweaks.

* Align cells across feature matrix tables

Make it easier for a user to follow a single column across multiple
tables of the feature matrix. Since the style sets cells to 25% width,
which assumes a table that is four cells wide, use a `style` tag so the
style only applies to the tables in the `teleport-editions.mdx` guide.

Closes gravitational/docs-website#134

* Update the MWI section of the feature matrix

Indicate that it is available for Community Edition users.

* Edit index.mdx

Apply changes made to identical language in teleport-editions.mdx.

* Respond to eglenn-teleport feedback

- Move the Community Edition column to the final place, rather than the
  first, since we now want to de-emphasize Community Edition and elevate
  the commercial offering.
- Edit the text to de-emphemphasize "editions", moving the description
  of editions to the end of the page and linking to this from the
  heading of relevant feature matrix columns.
- Remove section intro paragraphs, since the same information is
  on the home page of the docs.

* Add a Teleport Identity Security section



* Rename MWI features

---------

Co-authored-by: Ben Arent <ben@goteleport.com>