[v16] Move version support table to Upcoming Releases (#53154)
Backports #53046

This change helps guarantee that we'll update the version support table
at the same time as the upcoming release table to keep our support
windows up to date. Since we otherwise do not update the two tables at
the same time, the supported version table has become out of date.

This change also moves the compatibility description from FAQ to
Upcoming Releases so we keep this info in one place, and links to
Upcoming Releases from the relevant section of the FAQ.

Since the docs engine automatically copies the Upcoming Releases page
from the default version to all four versions of the docs, this change
does not edit the Upcoming Releases page for this version.