Fix 18 vale warnings (#53300)
Include inlining the partial:

docs/pages/includes/machine-id/recover-circle-ci-claims.mdx

This is because it's only used by two pages and, as a partial, violates
the rule in which each Var with a given name must be used multiple
times on a single page. Since this Vale rule uses the `raw` scope, it is
not possible to ignore it with an MDX comment.