[v17] Adds AWS Role ARN to IC Integration settings (#53034)
* [v17] Adds AWS Role ARN to IC Integration settings

Backports #53024

The Identity Center integration needs to be able to assume an AWS role
when using the ambient system credentials. This patch updates the IC
integration protobuf definition to include the role ARN, by refactoring
the settings block to replace the CredentialSource enumeration with a
`Credentials` type containing a `oneof` union of System and OIDC credential blocks.

The CredentalSource value is preserved for backwards compatibility with previous
Teleport releases.

This patch includes
- the updated protobuf message and generated types
- update tests, etc
- updated `tctl` to handle new plugin settings structure

* support awsic plugin --assume-role-arn flag

* align e-ref

---------

Co-authored-by: flyinghermit <sshah@goteleport.com>