Add tests to cover custom behavior in gravitational/go-oidc (#6264)
This adds tests to validate the changes introduced in
https://github.com/gravitational/go-oidc/pull/2,
https://github.com/gravitational/go-oidc/pull/4,
https://github.com/gravitational/go-oidc/pull/1/commits/ff27968050e30dbaccd9f1b31f4ea11ab1d89873.
The existence and continued passing of these tests should give us
confidence if and when we migrate off of our go-oidc to the
upstream version or to a different library entirely.

Note that only part of go-oidc/pull/1 is being covered here.
https://github.com/gravitational/go-oidc/pull/1/commits/25f9dc02f6d2b8251b514d4655f49052537a30e5
does not need to be carried forward.

Source-commit: gravitational/teleport.e@4f2f6e0bee792d65447579952b4515086e2a5373
Source-pull-requests: gravitational/teleport.e#6264