Auth: Rework resourceAccess to unifiedResourceLister (#53590)
There are no semantic changes whatsoever.

While I was working on support Okta read-only in Access Requests I needed to understand the way we
handle unified resources listing. While figuring this out I refactored the code so it is now easier
to comprehend to me. Hopefully it's also easier to comprehend to others.

This is a part of my branch with changes for Okta read-only support in Access Requests.