Integrate VNet info tab with first VNet launch and deep link (#53623)
* usePersistedState: Support function as arg to setState

* openExistingOrAddNew: Return doc URI

This is useful for further modifying the state of the doc that was just
opened.

* Adjust publicAddr and fqdn in app mock

This reflects the reality, as publicAddr and fqdn never include a port
number.

* Remove App reexport from tshd/types.ts

* Write test for first click on Connect next to TCP app

* Move more logic from connectToApp to useVnetLauncher

This is because we'll have to trigger different things depending on
whether VNet was launched for the first time or not. connectToApp does
not have access to that information.

* Rename VNet launcher to VNet app launcher

* Add state necessary to integrate info doc with first VNet launch

* Integrate info doc with first VNet launch

* Open info tab from VNet deep link

* Add missing contexts to TabHost story