Reduce resource consumption of (accessChecker) CheckAccess
Several of the helper functions called internally returned a helpful
debug message string that was included in logs. These strings were
allocated each time the function was called, leading to a significant
amount of memory consumption in clusters with a large number of
roles and or resources. Instead of crafting these debug messages
only for them to be later discarded, the helper functions now directly
log the messages at trace level.