Use checkboxes for resource permissions (verbs) (#53635)
* Use checkboxes for resource permissions (verbs)

As a prerequisite, the set of valid verbs is dynamically computed based
on the resource kinds picked for a given admin rule.

This change also narrows down the scope of compatible roles to those
that don't specify verbs that are irrelevant for a given set of resource
kinds.

Also turns action types supported by `useStandardModel()` into enums.
This makes it easier to navigate and refactor the code.

* review

* review