build: Add log groupings to release builds
Add log grouping to release builds of the builds themselves and the
buildboxes. On GitHub actions, this will add special log markers that
allow the log group to be collapsed, and when timestamps are enabled in
the log display, see how long these groups take to run. Building the
buildboxes when doing a release sometimes takes a while and emits a lot
of output, so this helps analyse the output. When running locally, a
markdown style heading is output.