[netiq] implement `refresh` token revocation (#6317)
* [netiq] implement `refresh` token revocation

This PR implements refresh token revocation for NetIQ tokens.
The NetIQ refresh tokens have a expiration of 5 minutes and must be
kept active otherwise the server forces a new login. For this purpose,
we do not reuse refresh tokens to refresh auth tokens because our
polling interval is 15m. This results in issuing a new login every
expiration/2 but until now we never actively revoked them which might
cause some issues with NetIQ.

To fix it, this PR introduces a token revocation mechanism that
invalidates old tokens once we login again or when the client is closed
- i.e. the process is shutting down.

Signed-off-by: Tiago Silva <tiago.silva@goteleport.com>

* Update lib/netiq/client/client.go

Co-authored-by: Zac Bergquist <zac.bergquist@goteleport.com>

---------

Signed-off-by: Tiago Silva <tiago.silva@goteleport.com>
Co-authored-by: Zac Bergquist <zac.bergquist@goteleport.com>

Source-commit: gravitational/teleport.e@6b366ae28ce13169ed6532e6921666f6d68e5c36
Source-pull-requests: gravitational/teleport.e#6317