fix: stop fetching SAML metadata in /webapi/ping (#53742)
`/webapi/ping` and `/web/config.js` will currently try to fetch SAML
metadata from an entity descriptor URL if one is configured. This can
break these endpoints if the IdP goes down or has network issues. To
make matters worse, we use the default HTTP client with no timeout on
the request.

This PR updates the ping and config.js to avoid fetching SAML metadata
over the network and adds a timeout to the request when we do fetch the
metadata.

Closes #44121