MWI: Extend max bot TTL to 7 days
This change extends the maximum allowed lifetime for bot certificates
to 7 days, up from the default maximum value of 24 hours. The default
values (1 hour TTL, 20 minute renewal) remain unchanged.

This somewhat circumvents the cluster's session TTL limit of 30
hours, however this limit is soft and can be overridden via roles
that specify an alternative MaxSessionTTL. This change takes
advantage of this behavior by setting this longer MaxSessionTTL on
the auto-generated bot roles, raising it from the previous default
of 12 hours.

As a result of this implementation detail, existing bots will not be
affected by this change, and will keep whichever MaxSessionTTL they
were created with, unless changed manually. Additionally, this can be
overridden with assigned bot roles as desired.