Add an EKS disaster recovery guide
When providing disaster recovery guidance for self-hosted Teleport
users, it is reasonable to begin with Amazon EKS users, since
self-hosted deployments on this platform are well supported. Provide a
guide for disaster recovery on EKS in the case of a regional outage.

This change adds the guide to the `helm-deployments` section, which
contains other guidance for self-hosted Teleport clusters on Kubernetes.
It changes the section index page to use `DocCardList` instead of an
incomplete, manually maintained list of guides.

While initial plans for this guide included a `tctl get all` discussion,
we don't recommend this approach because it doesn't back up all backend
data, so this change does not include this discussion in guide. Instead,
add a link to the Backup and Restore guide, which does discuss this
command.

Related changes:

- Extract DNS information in the `kubernetes-cluster.mdx` guide into a
  partial
- Fix an outdated DynamoDB docs link in the Backup and Restore guide