ci: Fix daily OCI image rebuild (#6350)
Define custom hybrid environments, `stage/oci-rebuild` and
`prod/oci-rebuild`, for running the
`rebuild-teleport-oci-distroless.yml` workflow. Currently it runs in the
`prod/build` environment, but that environment has OCI images going to
an internal pending registry that is immutable, and also signs the
images with pending keys. This fails because the registry/repo is
immutable.

The hybrid environment targets the public ECR registries and signs with
the public keys. The repos in the public registry are not immutable so
it can successfully push updated images.

Update `oci-image-build.yml` to use only the registry part of the
`ECR_PENDING_REPO_BASE` variable as when logging into `public.ecr.aws`,
it must not contain a repository suffix. In other cases it is ok, but
for this workflow to be able to push its results to the public registry,
this login change is needed.

Pass the new environment name from the top-level scheduled workflow
(`rebuild-teleport-oci-distroless-cron.yml`) for prod-based rebuilds.
The default for the called workflow is stage.

Source-commit: gravitational/teleport.e@46db556c1bbb02f49e3f44fca6eeea40e884c9c6
Source-pull-requests: gravitational/teleport.e#6350