Edit the Feature Matrix docs page (#53724)
* Edit the Feature Matrix docs page

Reflect the latest messaging guidance around Teleport products and
capabilities.

* Edit Feature Matrix formatting

- Expand the first column and narrow the final three columns.
- Use paragraphs for subsections of Teleport Zero Trust Access

* Add br tags before Feature Matrix paragraphs

This way, paragraphs that are supposed to divide the Zero Trust Access
table into subsections are easier to pick out from the tables they
divide. We can use this approach until we introduce a plugin to allow
colspans.

* Respond to roraback feedback

- Fix empty SSO cell
- Fix empty "Machines" cells
- Fix empty Access Requests cell
- Remove "Teleport" from table headers

* Include table subsections as shaded rows

Edit the Feature Matrix page to:
- Shorten table section descriptions
- Use a style tag to shade certain rows of the first and final tables

* Fix spelling

* Fix absolute links in the Feature Matrix

* Add missing feature matrix features

Add two items to the Identity Governance section