Stabilize TestInteractiveCompatibilityFlags (#54007)
The test was immediately attempting to connect to the target host,
but due to cache propagation the proxy may not yet have it's
inventory updated. This sporadically causes the test to fail as
dial requests to the target cannot be honored. To make the test
more robust to this situation a retry mechanism was added. In other
tests we first wait for the node to be visible in the correct cache,
however, this is not exposed in the TeleportProcess and thus not
available to this test.

Closes #53736.

Co-authored-by: Zac Bergquist <zac.bergquist@goteleport.com>