fix: handle switching between VNet apps on Windows (#54010)
Currently VNet Windows is pretty broken, if you connect to 2 different
apps in the same cluster and try to switch back and forth between them,
you will get TLS handshake errors and app connections will not work.
This is because:
- on Windows the VNet service has no access to user private keys and
  uses a crypto.Signer implemented over a gRPC service provided by tshd
- clientApplicationService holds a map of crypto.Signers for each app
  keyed by AppKey
- localAppProvider left AppKey.Name empty, so all AppKeys in a cluster
  were matching and every new app cert issued would overwrite the signer
  used by all other apps in the cluster

This commit:
- updates TestRemoteAppProvider to catch the bug
- updates localAppProvider to correctly fill in AppKey.Name
- adds some sanity checking of the app key to clientApplicationService

Changelog: Fixed TLS errors when switching between VNet apps on Windows.