[v17] MWI: Extend max bot TTL to 7 days (#53893) (#54063)
Backport of #53893 to branch/v17

---

This change extends the maximum allowed lifetime for bot certificates
to 7 days, up from the default maximum value of 24 hours. The default
values (1 hour TTL, 20 minute renewal) remain unchanged.

This somewhat circumvents the cluster's session TTL limit of 30
hours, however this limit is soft and can be overridden via roles
that specify an alternative MaxSessionTTL. This change takes
advantage of this behavior by setting this longer MaxSessionTTL on
the auto-generated bot roles, raising it from the previous default
of 12 hours.

As a result of this implementation detail, existing bots will not be
affected by this change, and will keep whichever MaxSessionTTL they
were created with, unless changed manually. Additionally, this can be
overridden with assigned bot roles as desired.

---

* Make max_session_ttl configurable via Bot resource

* Fix tests; fix inclusion of field in generated bot specs; fix update

* Update operator CRDs

* Attempt to fix integration test

* Update docstring with sane units

* Update crdgen test data

* Fix failing tests by enabling UseProtoNames

Also fixes this in the TeleportWorkloadIdentityV1 resource, which was
affected by the same problem.