Web: Use `$HOME/.config/teleport` folder for App Access
Instead of using `$HOME/.config` folder, which is the
XDG_CONFIG_HOME variable, used by a lot of applications to store their
configuration.

If teleport owns this directory, other applications will misbehave and
we are not complying with the freedesktop specification
https://specifications.freedesktop.org/basedir-spec/latest/

When this folder is not empty, teleport emits a warning:

```
WARNING: The data directory /root/.config is not empty and may contain existing cluster state. Running this configuration is likely a mistake. To join a new cluster, specify an alternate --data-dir or clear the /root/.config directory.
```