S3: add a configurable option to ignore some uploads (#54201)
Teleport's uploader was written under the assumption that Teleport
is the only thing generating uploads in the configured bucket.
If this is no longer true, Teleport will attempt to complete uploads
that it didn't start, potentially truncating the upload prematurely.

This hasn't been a problem to date, but new efforts to enable
replication across regional buckets are one use case where our
previous assumptions break.

This change allows us to specify an allow list of initiators to complete
uploads for, ensuring that Teleport won't stomp on or attempt to process
other uploads in any way.