Change the size of "add more" buttons to small
This change is motivated by the role editor UI, where we want to
introduce a visual hierarchy of buttons (large - dialog actions, medium
- adding sections, small - actions inside sections). Note that this
could also be configurable, but it doesn't look bad in the other place
where the labels input control is used, so I decided to trim the
complexity. We may revisit this when we decide to unify the label inputs
across the role editor and the Discovery feature.