Fix avatar link in the Discord guide (#54312)
The avatar image download link in the Discord guide 404s. Use the
approach we took in #51193 to fix the link: include an `<a>` tag instead
of a Markdown link, and use an `import` statement to retrieve the URL
path of the static asset. The URL path points to the `static` directory,
where we host images reachable by import statements in Docusaurus.