fix: fixes a possible panic during auth start if upsert operation fails (#54326)
This PR fixes a possible panic caused by a failure caused by a backend
failure. If the failure occurs, we replaced the user object with the
result of the `{Upsert|Create}User` operation. If the operation fails,
both calls return `nil, err` and when building the error stream, we
accessed a nil user.

Signed-off-by: Tiago Silva <tiago.silva@goteleport.com>