MWI: Bound Keypair Joining - Minimal Client Implementation
This contains a minimal client-side implementation of bound keypair
joining for `tbot`. This first iteration supports only preregistered
keys, and so adds a new `tbot keypair create` subcommand that
generates a keypair and prints a public key that can be copied into
a token.

In an effort to still take advantange of tbot's secure filesystem
abstraction and to not permanently tie the new join method to it,
this includes a minimal filesystem wrapper to adapt tbot's
destinations to a generic file read/write interface.