Convert security reports to new in memory cache collection
Moves audit queries, security reports and security report states
to the new cache collection scheme that was introduced in #52210.
This also converts the new collections to correctly fetch the
resources from the actual upstream, cleans up extraneous methods
from the SecReports interface and removes the last of the
legacy cache collection machinery since it is no longer used.

Closes #54520.