[v17] Access Graph Demo Mode (#54290)
* Add DemoMode and Status to AccessGraphSettings

* Add RoleEditorVisualizer to Role Editor (#53535)

* Allow cloud to update access graph settings (#53800)

* Move demo mode check to RoleEditor (#53961)

* Add optional chaining to demo mode check in RoleEditor (#53969)

* Update proxy AccessGraphSettings permissions (#54009)

This allows proxy to Read AccessGraphSettings, as well as prevents the
"Preview Identity Security" button from showing up if the user does not
have access to update access graph settings

* Add AccessGraphDemoMode entitlement

* Add DemoMode entitlement to Identity Security Preview button (#54300)

This adds one extra conditional to display the access graph demo button.
The user must also have the entitlement enabled. This is already checked
when enabling demo mode in the backend, but with this frontend check we
will hide the button if the request was going to fail anyway.