Fix 31 Vale warnings and one error
Fix step numbeing and `Var` name mismatch issues in the Azure Sync
guide.

Fix minor issues in the following guides:
- AWS Console
- Azure AKS Workload ID
- Google Cloud API guide

Add "How it works" H2s to pages that already include an architectural
explanation at the top of the page, and only need the H2 to be
consistent with the intended structure of a docs how-to guide.