Fix cache listing access list members (#54528)
The `(Cache) ListAccessListMembers` implementation did not properly
specify an end key to the resource range which resulted include
members for all access lists for the desired list and any with a
name between it and the end of the range.

Additionally, the unused simple service that the cache used to use
for access lists was removed.