Replace graviational/go-oidc with zitadel/oidc (#6383)
Our outdated and unmaintained go-oidc fork made it challenging to
introduce new features and required workarounds to bugs or oddities
with the forked version of go-oidc.

The choice of zitadel/oidc over moving to the upstream coreos/go-oidc
library was motivated by a few different things:

1) The zitadel library is certificed by OpenID foundation
2) It supports PKCE out of the box
3) It supports a server implementation which will unlock
   https://github.com/gravitational/teleport/issues/20967

While most of the functionality was fairly straight forward to
migrate didn't require additional consideration there are
a few Teleport-specific items that required work arounds. While the
library does allow specifying a custom HTTP client, it doesn't
expose internals of each HTTP call. This made our existing logic
which rejects discovery responses with a large response size and
handling of a subset of userinfo response codes challenging. In
order to preserve the existing behavior custom round trippers were
put in place to hook into the HTTP requests.

The new library also eliminates several classes of bugs and special
behavior that had been crafted over the years to tolerate go-oidc.
Additionally, we've received confirmation that we do not need to
carry forward some of the NetIQ specific hacks that had been included
in our go-oidc fork.

The code and tests were also restructured to eliminate injecting
claims and bypassing communication to the idp for tests. The testing
suite was rewritten to launch an oidc server and all tests now
fully exercise end to end communication with said server instead of
manually injecting claims. This helps to ensure that tests are
exercising code paths as close to production workflows as possible.
Additionally, a few tests that no longer applied were removed, and
`TestOIDCAuthRequest` is being moved to OSS.

Source-commit: gravitational/teleport.e@86ae4a521980de9f3c394a80f236d9fc4ebd070d
Source-pull-requests: gravitational/teleport.e#6383