Refactor `enroll-resources/agents`
Give the section a more reasonable information architecture. The section
includes two children, one of which is a subsection (Joining Services)
and another is an introductory overview.

Move `agents/introduction.mdx` to make it the new `agents/agents.mdx`
section index page. This page is a high-level introduction, so it makes
more sense as the section index. Also edit the page to:

- Remove an unnecessary list of links to sections for each agent
  service.
- Condense the architecture overview H2 to pave the way for making this
  page more of a section intro and less of an intro/conceptual hybrid
  guide.

Move "Join services" guides to make them direct descendents of the
`agents` directory. This way, the substantial content of the section can
have a higher directory level.

Extract the information on adding a service to an existing agent into a
full how-to guide. This makes the instructions more visible. Also
include example snippets within the guide.