Fix test types for a wrapper of Button
I couldn't get it to work when destructuring rest props, e.g.:

    const { size, children, ...buttonProps } = props;

so for now I just removed it altogether. I don't know why it works
in ForwardedAsWrapper, I suppose it has to do either with the css prop
passed to ButtonPrimary? Or maybe the forwardedAs prop accepted by
ForwardedAsWrapper?