ci: Update CI buildbox to teleport18
Update the buildbox used by the various CI jobs to `:teleport18` from
`:teleport17`. This should have been done when `branch/v17` was created,
but got missed.

The `unit-test-*` have been left out as there seems to be an issue with
some workflow runs erroring out with "The request could not be processed
because too many files changed". I'll raise a second PR for those.