Convert headless authentication to new cache mechanism
Moves headless authentication resources to the new cache collection
scheme that was introduced in #52210. This collection is a bit
different than the rest since the cache does not need to store the
resources, the collection exists solely so that watchers of the
headless authentication types can be created.

In the future we should consider refactoring the machinery for types
that exist solely  for watchers, as of now there are only two:
headless authenticaion and access requests.