Field naming refactor
This refactors a number of bound keypair terms to simplify the token
and improve clarity.

We now consistently use these terms to refer to various concepts in
the join method:
- "onboarding" to refer to the first join
- "registration" to refer to the binding of a keypair to a token
- "preregistration" to refer to keys added manually with the token
- "registration secret" to refer to keys bound at the time of
  onboarding using a secret token
- "recovery" to refer to joining with no or an expired identity

Various fields were renamed to reflect these new terms and
definitions.

Additionally, the "unlimited" and "insecure" fields have been merged
into an enum-style string, with values "standard", "relaxed", and
"insecure".