Add codeowners for /web dep updates
With the dependabot.yaml deprecation the suggested change is to add
reviewers to the CODEOWNERS file directly. Unfortunately this doesn't
seem to give us the granularity to select "only dependabot PRs" and so I
think the next best thing is just assign us to any PR that updates the
package.json files (which is usually just dependabot).