[v17] ci: Pre-install wasm-bindgen on macos CI build
Pre-install the correct version of `wasm-bindgen` in the
`build-macos.yaml` workflow when we install `wasm-pack`.

If `wasm-bindgen` is not installed when `wasm-pack` is run, it will
build it itself but do so without `--locked`. This ends up with "random"
versions of dependencies, some of which require a newer version of rust
than we have installed.