ci: Pre-install wasm-bindgen on macos release build (#6506)
Pre-install the correct version of `wasm-bindgen` in the
`build-macos.yaml` workflow when we install `wasm-pack`.

If `wasm-bindgen` is not installed when `wasm-pack` is run, it will
build it itself but do so without `--locked`. This ends up with "random"
versions of dependencies, some of which require a newer version of rust
than we have installed.

Source-commit: gravitational/teleport.e@e61ea9224f01e41c8b583a4b2ace8e7e13278714
Source-pull-requests: gravitational/teleport.e#6506