Move auth preference module validation to RPC layer (#54687)
The module validation rejects auth preferences that have second
factor disabled without the environment variable override. Doing
this in the storage layer means that in order to disable second
factor the environment variable needs to be set on _all_ teleport
processes not just Auth. This can result in caches of downstream
agents from becoming healthy until the manual override is applied.
The intent is to prevent modifying an the auth preference to disable
second factor, which when moved to the RPC layer, has the same
affect without the possibility of caches performing extra validation.