WMI: Add `tbot/service` package
Introduces a new set of abstractions for working with the different components
of tbot, making them loosely-coupled such that if a service you depend on fails,
you can timeout and degrade gracefully.

Example:

    type MyService struct {
        Dependency *service.Service[Dependency]
    }

    func (s *MyService) Run(ctx context.Context, _ *service.Runtime) error {
        ctx, cancel := context.WithTimeout(ctx, 2*time.Second)
        defer cancel()

        // Wait for the dependency to become healthy.
        dep, err := s.Dependency.Wait(ctx)
        if err == nil {
            // Use the dependency.
        } else {
            // Degrade gracefully.
        }
    }

This lays the groundwork for allowing services to run with cached credentials
when the auth server is unavailable, which is currently impossible because if
the bot identity service initialization (i.e. renewing the identity) fails we
terminate `tbot` immediately.